freedreno/ir3: add support for ssbo intrinsics
The non-inc/dec encoding can use some work. But for now that's all we
get in practice.

We add an extra fake register for the benefit of instruction ordering,
based on the last memory op. This could be made smarter to work on a
per-buffer or even per-address basis, but the current solution is
simple.

Signed-off-by: Ilia Mirkin <imirkin@alum.mit.edu>