mesa/copyimage: allow width/height to not be multiples of block
For compressed textures, the image size is not necessarily a multiple of
the block size (e.g. the last mip levels). For example the RGTC spec
calls out this condition as legal, saying to error only when:

    * <width> is not a multiple of four, and <width> plus <xoffset>
    * is not equal to TEXTURE_WIDTH;

While the GL_ARB_copy_image spec does not call this out explicitly, it
appears that some games rely on this. Also it seems like it should be
possible to copy in the last miplevels of a compressed texture's
miptree.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=92860
Signed-off-by: Ilia Mirkin <imirkin@alum.mit.edu>
Cc: mesa-stable@lists.freedesktop.org