nv50,nvc0: fix frag coord when half-pixel center is disabled
The frag coord goes with the half pixel center setting. Since this is a
roughly global setting, it's safe to avoid creating shader variants.
When it is disabled, we add (0.5,0.5) to the frag coord to make it be in
the center as advertised by our caps. This only happens for st/nine, as
otherwise it is always flipped in.

Signed-off-by: Ilia Mirkin <imirkin@alum.mit.edu>