nv50/ir: don't rely on inbound edge order in phi nodes
Previously, a phi node's sources were implicitly ordered by the inbound edge
order. This patch changes that so that a phi node instead has a basic block
stored for each source in a deque.

There are no regressions in Unigine Heaven, Valley and Superposition (with
a SERIALIZE hack to fix unrelated errors in all three).

I also found that the code generated is the same as without this patch for
the shaders in shader-db even with the sources of phi nodes reversed after
convertToSSA().